feat: add screenshots to documentation
This commit adds relevant screenshots to various sections of the documentation. The purpose of this addition is to provide users with visual aids that can promote better understanding of the content and usage of different sections in the software. The screenshots demonstrate key interfaces and features which may help users navigate through the software more easily.

Asana: https://app.asana.com/0/1203851531040615/1206090378779159/f